Implement ScriptRunner Interface that uses Java's ServiceLoader mechanism.
The Jython ScriptRunner is contained in the SikuliX-Jython-API project.
This new layout separates Jython from Sikuli-API to allow an easier creation and integragion of ScriptRunners.
With this update, Jython/Python scripts can only be executed if the .jar that can be created from the SikuliX-Jython-API project is in the classpath or the .jar of the SikuliX-Jython-API with includes is used.
Beware: SikuliX-Jython-API project is in development and while commandline execution of scripts with -r and interactive sessions with -i are working, the integration with Sikuli-IDE might be broken.
To fix this, Sikuli-IDE probably has to be updated so that it can use the IScriptRunner interface as well.

Conflicts:

	pom.xml -> small update to remove an unneeded definition
	src/main/java/org/sikuli/script/Region.java -> tabs replaced with whitespaces and fixed indent, also removed some Python-specific code
	src/main/java/org/sikuli/script/SikuliScriptRunner.java -> deleted, is now implemented in SikuliX-Jython-API